feat: debugging
