+++ index.php
@@ -2,6 +2,6 @@
 // index.php
 
 // Redirect to login.php
-header("Location: login.php");
+header("Location: /src/pages/login.php");
 exit; // Ensure that no further code is executed after the redirection
 ?>
